Create base map component for obs and site maps
